Deleted the progams function for now and implemented the backend for creating workouts
And started on the fronted for the workouts functionallity

+++ backend/src/controllers/exercise.controller.js
@@ -1,51 +1,96 @@
 import Exercise from "../models/models.exercise.js"
 import mongoose from "mongoose"
 
+/*Backend API for exercise.model
+Exercise Controller made to manage the MONGODB database of exercises
+ getExercises --> get all the exercises stored in the database
+ createExercise --> Create a new exercise in the database
+ deleteExercise --> Delete an exercise id from the database
+ updateExercise --> Update the value of a exercise
+ */
+
+// A function to get all the exercises from the database
 export const getExercises = async (req, res) =>{
 
     try{
-        const exercise = await Exercise.find({})
-        res.status(200).json({success: true, data: exercise})
+        //Mongoose finds all the exercises
+        const exercises = await Exercise.find({})
+        //returns an array of all the found exercises as JSON
+        res.status(200).json({
+            success: true, 
+            data: exercises
+        })
     }catch(error){
         console.log("Error in getExercises", error.message)
-        res.status(500).json({success: false, message: "Server error"})
+        res.status(500).json({
+            success: false, 
+            message: "Server error"})
 
     }
 }
 export const createExercise = async (req, res) =>{
-    
+    //The exercise sent as JSON from frontend
     const exercise = req.body;
 
+    //Simple validation
     if(!exercise.name || !exercise.description){
-        return res.status(400).json({success:false, message: "Fill in all fields"})
+        return res.status(400).json({
+            success:false,
+            message: "Fill in all fields"
+        })
     }
-
+    //Creates the exercise with mongoose
     const newExercise = new Exercise({
-        ...exercise,
-        //If it is created by a user it is a user specific exercise
-        createdBy: req.user._id 
+        name: exercise.name,
+        description: exercise.description
     })
 
     try{
+        //Saves the exercise in the database
         newExercise.save()
-        res.status(201).json({success: true, data: newExercise})
+        //Responds with the added exercise as JSON (+success: true)
+        res.status(201).json({
+            success: true, 
+            data: newExercise
+        })
 
     }catch(error){
         console.log("Error in createExercise", error.message)
-        res.status(500).json({success: false, message: "Server Error"})
+        res.status(500).json({
+            success: false, 
+            message: "Server Error"
+        })
 
     }
 }
 export const deleteExercise = async (req, res) =>{
+    //extract the id from the req URL parameter
     const {id} = req.params
-    console.log("Deleting", id)
+
+    //Checks that the provided id is valid MongDB ID
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({
+            success: false, 
+            message: "Invalid Exercise Id"
+        })
+    }
     
     try{
+        //Try to find and delete exercise
         await Exercise.findByIdAndDelete(id)
-        res.status(200).json({success: true, message: "Exersice deleted"})
+
+        //if success: 
+        res.status(200).json({
+            success: true,
+            message: "Exersice deleted"
+        })
+        //if server down or other error:
     }catch(error){
         console.log("Error in deleteExercise", error.message)
-        res.status(500).json({success: false, message: "Server Error"})
+        res.status(500).json({
+            success: false, 
+            message: "Server Error"
+        })
     }
 }
 
@@ -55,13 +100,24 @@ export const updateExercise = async (req, res) =>{
     const exercise = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)){
-        return res.status(404).json({success: false, message: "Invalid Exercise Id"})
+        return res.status(404).json({
+            success: false, 
+            message: "Invalid Exercise Id"
+        })
     }
     try{
         const updatedExercise = await Exercise.findByIdAndUpdate(id, exercise, {new:true})
-        res.status(200).json({success: true, message: "Exercise updated", data: updatedExercise})
+        
+        res.status(200).json({
+            success: true, 
+            message: "Exercise updated", 
+            data: updatedExercise
+        })
     }catch(error){
         console.error("Error in updateExercise", error.message)
-        res.status(500).json({success: false, message: "Server Error"})
+        res.status(500).json({
+            success: false,
+            message: "Server Error"
+        })
     }
 }
